requirements.txt
python-dotenv==1.1.1
google-genai==1.36.0
PyPDF2==3.0.1
redis==6.4.0
requests==2.32.5
Flask==3.1.2
flask-cors==6.0.1